Restructure library a little bit
