+++ src/lib.rs
@@ -2,6 +2,8 @@
 extern crate cpython;
 
 mod darken;
-mod darken_py;
+mod errors;
+mod python;
 
-pub use darken::{darken_pixels, ImgError};
+pub use darken::darken_pixels;
+pub use errors::ImgError;
